+++ user/src/main/java/com/pch/user/service/CommodityService.java
@@ -1,22 +1,21 @@
 package com.pch.user.service;
 
-import java.util.List;
-
 import com.pch.user.dto.CommodityDto;
+import java.util.List;
 
 public interface CommodityService {
 
-	/**
-	 * 查询所有商品
-	 *
-	 * @return 商品集合
-	 */
-	List<CommodityDto> queryAllCommodity();
+    /**
+     * 查询所有商品
+     *
+     * @return 商品集合
+     */
+    List<CommodityDto> queryAllCommodity();
 
-	/**
-	 * 添加商品
-	 *
-	 * @return 成功 > 0
-	 */
-	int insertCommodity(CommodityDto commodityDto);
+    /**
+     * 添加商品
+     *
+     * @return 成功 > 0
+     */
+    Long insertCommodity(CommodityDto commodityDto);
 }
